BencherServer/pyproject.toml
[project]
name = "bencherserver"
version = "0.1.0"
description = ""
authors = [{ name = "Leonard Papenmeier", email = "leonard.papenmeier@gmail.com" }]
requires-python = "~=3.11.0"
readme = "README.md"
dependencies = [
    "grpcio>=1.60.1,<2",
    "protobuf>=4.25.2,<5",
    "bencherscaffold>=0.3.1,<0.4",
]

[project.scripts]
start-benchmark-service = "bencherserver.main:serve"

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"
